requirements.txt
opencv-python~=4.10.0.84
matplotlib~=3.7.5
PyQt6~=6.7.1
numpy~=1.24.4
transliterate~=1.10.2
timm~=1.0.15
torch~=2.4.1+cu118
torchvision~=0.19.1
DateTime~=5.5
pillow~=10.3.0
scenedetect~=0.6.6
PySide6~=6.6.3.1